+++ core/setup/systemd/clean_service.sh
@@ -6,11 +6,10 @@ if [ "$EUID" -ne 0 ]; then
 fi
 
 sudo rm /etc/systemd/system/cyclosafed.service
-sudo rm /etc/systemd/system/pigpiod.service
 sudo rm /etc/systemd/system/gpiod.service
 sudo rm /etc/systemd/system/gps_time.service
 
-systemctl stop cyclosafed.service gps_time.service pigpiod.service gpiod.service
-systemctl disable cyclosafed.service gps_time.service pigpiod.service gpiod.service
+systemctl stop cyclosafed.service gps_time.service gpiod.service
+systemctl disable cyclosafed.service gps_time.service gpiod.service
 systemctl daemon-reload
 systemctl reset-failed